More inputs

+++ inputs/grades.py
@@ -1,10 +1,10 @@
 """Write a python code to take input of marks in 5 subjects and print the total and the average."""
 
-math_marks=int(input("Type your marks in math:"))
-english_marks=int(input("Type your marks in english:"))
-gym_marks=int(input("Type your marks in gym:"))
-science_marks=int(input("Type your marks in science:"))
-art_marks=int(input("Type your marks in art:"))
+math_marks=float(input("Type your marks in math:"))
+english_marks=float(input("Type your marks in english:"))
+gym_marks=float(input("Type your marks in gym:"))
+science_marks=float(input("Type your marks in science:"))
+art_marks=float(input("Type your marks in art:"))
 total_marks=math_marks+english_marks+gym_marks+science_marks+art_marks
 avg_marks=total_marks/5
 
@@ -14,3 +14,6 @@ if avg_marks>75:
     print("You passed!")
 else:
     print("You failed!")
+
+
+
